% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemann.R
\name{riemann}
\alias{riemann}
\title{Riemann integration}
\usage{
riemann(f, range, n)
}
\arguments{
\item{f}{a function with a single argument x}

\item{range}{a vector specifying the lower and upper end points of the range of integration (a and b)}

\item{n}{the number of intervals for the Riemann approximation}
}
\value{
a single value which is the Riemann approximation to the integral
}
\description{
Riemann integration
}
\examples{
> riemann(function(x){x^2}, c(1,2), 100)
2.34835
}
