% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpsons.R
\name{simpsons}
\alias{simpsons}
\title{Simpson's Rule}
\usage{
simpsons(f, range, n)
}
\arguments{
\item{f}{a function with a single argument x}

\item{range}{a vector specifying the lower and upper end points of the range of integration (a and b)}

\item{n}{the number of intervals for the Simpsons approximation}
}
\value{
a single value which is the Simpsons approximation to the integral
}
\description{
Simpson's Rule
}
\examples{
simpsons(function(x){x^2},c(1,20),500)
2666.333
}
